Instead of raising UnsupportedAuthenticationError on tss, use an allowlist of known working sources
